+++ demo/molecule/autosuggest/demo/index.js
@@ -12,6 +12,8 @@ import withDynamicOptions from './hoc/withDynamicOptions'
 
 import AutosuggestSingleWithAsyncOptions from './components/AutosuggestSingleFromAjax'
 
+import ComboCountries from './components/ComboCountries'
+
 import {IconClose} from './Icons'
 import {getAsyncCountriesFromQuery} from './services'
 import './index.scss'
@@ -70,6 +72,16 @@ const Demo = () => (
       />
     </div>
 
+    <div className={CLASS_DEMO_SECTION}>
+      <h3>disabled</h3>
+      <MoleculeAutosuggestWithState
+        value="Luxembourg"
+        onChange={(_, {value}) => console.log(value)}
+        iconClear={<IconClose />}
+        disabled
+      />
+    </div>
+
     <h2>Multiple Selection</h2>
     <p>
       Este componente permite añadir nuevas opciones (como tags) aunque no esten
@@ -95,6 +107,24 @@ const Demo = () => (
         iconClear={<IconClose />}
         multiselection
       />
+    </div>
+
+    <div className={CLASS_DEMO_SECTION}>
+      <h3>Disabled</h3>
+      <MoleculeAutosuggestWithStateTags
+        tags={['India', 'Luxembourg']}
+        onChangeTags={(_, {tags}) => console.log(tags)}
+        iconCloseTag={<IconClose />}
+        iconClear={<IconClose />}
+        multiselection
+        disabled
+      />
+    </div>
+
+    <h2>Dependant Selection</h2>
+    <div className={CLASS_DEMO_SECTION}>
+      <h3>With Placeholder</h3>
+      <ComboCountries />
     </div>
   </div>
 )