+++ components/atom/spinner/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.6.0"></a>
+# 1.6.0 (2019-09-02)
+
+
+### Features
+
+* hooks migration ([cbe4ccc](https://github.com/SUI-Components/sui-components/commit/cbe4ccc))
+
+
+
 <a name="1.5.0"></a>
 # 1.5.0 (2019-06-26)
 
@@ -41,6 +51,24 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 
 * fix z index and change demo example ([04e205b](https://github.com/SUI-Components/sui-components/commit/04e205b))
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2018-03-07)
+
+
+### Bug Fixes
+
+* pull request changes ([ed237f9](https://github.com/SUI-Components/sui-components/commit/ed237f9))
+
+
+### Features
+
+* delayed loading + playground examples ([01be8a8](https://github.com/SUI-Components/sui-components/commit/01be8a8))
+* new atom spinner component ([0e818e1](https://github.com/SUI-Components/sui-components/commit/0e818e1))
+* placeholder and alignment ([00dd457](https://github.com/SUI-Components/sui-components/commit/00dd457))
+* removing show prop ([8e411d0](https://github.com/SUI-Components/sui-components/commit/8e411d0))